+++ src/components/__tests__/Page.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { render, screen } from "utils/utils.helper.tsx";
+import { render, screen } from "utils/utils.helper";
 import Page from "../Page";
 
 it("should render without crashing", () => {