+++ src/global.d.ts
@@ -1,3 +1,5 @@
-// declare global values here, which are available throughout the codebase.
-// these get set via Vitejs, within the vite.config.ts file
+/*
+ * Declare global values here, which are available throughout the codebase.
+ * these get set via Vitejs, within the vite.config.ts file
+ */
 declare const APP_VERSION: string;