+++ src/utils/api/api.ts
@@ -1,4 +1,4 @@
-// a super simple (not finished) wrapper for NXT API calls
+// A super simple (not finished) wrapper for NXT API calls
 
 export interface IAPIResult {
   notsure: string;
@@ -15,7 +15,7 @@ export async function API<IAPIResult>(
       "Content-Type": "application/json",
       "Access-Control-Allow-Origin": "*",
     },
-    method: method,
+    method,
     body: JSON.stringify(data),
   });
   const jsonResult = await result.json();