+++ src/views/Login/components/OnboardingStepper/components/1.GenerateSeedStep/GenerateSeedStep.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback } from "react";
+import React, { memo, useCallback } from "react";
 import Button from "@mui/material/Button";
 import useAccount from "hooks/useAccount";
 import { IStepProps } from "../types";
@@ -6,14 +6,13 @@ import { styled, Typography } from "@mui/material";
 
 const GenerateSeedStep: React.FC<IStepProps> = ({ stepForwardFn }) => {
   const { fetchFn } = useAccount();
-
-  // fetches a fresh wallet and then steps the stepper to the next step
+  // Fetches a fresh wallet and then steps the stepper to the next step
   const handleGenerateClick = useCallback(async () => {
     if (fetchFn === undefined) {
       return;
     }
 
-    // await because we need accountRs and accountSeed to be populated before moving forward (or the next step will throw alert)
+    // Await because we need accountRs and accountSeed to be populated before moving forward (or the next step will throw alert)
     await fetchFn(); // TODO: re-consider the way this is handled, does it make more sense to check on the next step?
     stepForwardFn();
   }, [fetchFn, stepForwardFn]);
@@ -41,4 +40,4 @@ const StyledButton = styled(Button)(({ theme }) => ({
   },
 }));
 
-export default React.memo(GenerateSeedStep);
+export default memo(GenerateSeedStep);